+++ griptape/loaders/__init__.py
@@ -6,7 +6,6 @@ from .web_loader import WebLoader
 from .sql_loader import SqlLoader
 from .csv_loader import CsvLoader
 from .dataframe_loader import DataFrameLoader
-from .file_loader import FileLoader
 from .email_loader import EmailLoader
 from .image_loader import ImageLoader
 
@@ -20,7 +19,6 @@ __all__ = [
     "SqlLoader",
     "CsvLoader",
     "DataFrameLoader",
-    "FileLoader",
     "EmailLoader",
     "ImageLoader",
 ]